Impl: icons with support for light&dark themes (#32)
- LAF support in Toolbox is quite primitive, it turns out icon support
for light and dark themes is enabled by a masked flag on the icons.
- the mask flag controls whether the svg colors are inverted in
light&dark themes.

Among other things we also fixed and issue with the `Start` button which
remained active when a workspace was stopped and outdated. In order to
be more consistent with the web client we renamed the button to `Update
and start` to reflect that it also starts the workspace.

Running and outdated workspaces also received a new action button:
`Update and restart`.

- resolves #31

![image](https://github.com/user-attachments/assets/f54a71f4-ca8d-4655-b5ff-e95c52e3ad9a)

![image](https://github.com/user-attachments/assets/a81901ac-b7fe-482d-bf09-13baaa02455b)