Impl: update ssh config when settings change (#51)
- right now the ssh config is triggered only when at the login and only
when new workspaces are
created. In the last case, the ssh config updates only the new sections.
- but the user can go into settings page and enable or disable the
wildcard ssh config. With this patch
after the user hits Save, the ssh re-configuration is triggered, without
the need to restart Toolbox.

Also a major fix - read only settings acted as snapshots in the sense
that they did not reflect any subsequent update to the underlying store
because updates create a new readonly instance
- with this patch we simplified the code even more by exposing a
readonly interface with all of the
implementation in the CoderSettingsStore. PluginSettingsStore and
Environments are the only persist-able
  stores.
- because read only instances share the same settings store instances
any update on the writable will reflect
  on the readable instances as well.