fix: token input screen is closed after switching between Toolbox and browser (#72)
- rough draft to fix UI state management in the authentication flow
which today has 3 pages. If user closes
Toolbox in any of these three pages (for example to go and copy the
token from a browser), then when it comes back in Toolbox does not
remember which was the last visible UiPage.
- until JetBrains improves Toolbox state management, we can work around
the problem by having only one
UiPage with three "steps" in it, similar to a wizard. With this approach
we can have complete control over
  the state of the page.

- to be noted that I've also looked over two other approaches. The first
idea was to manage the stat
ourselves, but that didn’t work out as Toolbox doesn’t clearly tell us
when the user clicks the Back
button vs. when they close the window. So we can’t reliably figure out
which page to show when it reopens.
- another option was changing the auth flow entirely and adding custom
redirect URLs for Toolbox plugins.
But that would only work with certain Coder versions, which might not be
ideal.

- resolves #45