fix: rendering glitches when a Workspace is stopped while an SSH connection is alive
Toolbox raises a class cast exception when Workspaces are stopped while the SSH connection
is running. After the workspace was stopped Toolbox refused to show widget with some weird glitches
on the screen.
The fix in this case is to safely disconnect the SSH before sending the stop command to the workspace.
The code will wait at most 10 seconds for the disconnect to happen, and only after that send the stop.