fix: access the settings page for the auth. wizard
The Settings menu is only available after we successfully authenticate.
This can be somewhat problematic if we want to configure something like
coder cli path, or certificate path before doing the authentication.

A new "Settings" button at the bottom of the page can now access the Settings
page. Toolbox is a bit inflexible with the API because:
- I could not find a way to delimit or separate the Settings button from the Back and Sign In/Connect button
- we can't put a button or anything else in the top right corner, the traditional place for a settings icon.

- resolves #90