impl: identify the PID for the SSH command
Toolbox spawns a native process running the SSH client. The ssh client then spawns
another process which associated to the coder proxy command. SSH network metrics
are saved into json file with the name equal to the pid of the ssh command (not
to be confused with the proxy command's name)