impl: network metrics poll job
When ssh is connected a poll job is created that searches for the
SSH command pid, and then determines the network metrics json file.

The pid can change multiple times in a Toolbox session, especially
when the OS goes to sleep and comes back - the ssh is respawned.
This is the reason we don't cache the pid, instead always search
for it.

The poll job activates every 5 settings and it is cancelled when
the user stops the ssh connection.