impl: support for displaying network latency (#108)
Under the "Additional environment information". Unfortunately it was not
possible any other way. The description property is modifiable however
Toolbox renders the description label only as long as the SSH connection
is not established. As soon as an ssh connection is running the
description label is used as mechanism to notify users about available
IDE updates.

It also appears that we can't have any other extra tab, other than
"Tools", "Projects" and "Settings". There is a secondary information
attribute API, but it is not usable to show recurring metrics info
because it can only be
configured once, it is not a mutable field.

The best effort was to add the information in the Settings page, and it
is worth highlighting that the metrics are only refreshed when user
either:
- switches between tabs
- expands/collapses the "Additional environment information" section.

There is no programmatic mechanism to notify the information in the
Settings page that latency changed.

The network metrics are loaded from the pid files created by the ssh
command. Toolbox spawns a native process running the SSH client. The ssh
client then spawns another process which is associated to the coder
proxy command. SSH network metrics are saved into json files with the
name equal to the pid of the ssh command (not to be confused with the
proxy command's name).

- resolves #100 
- resolves #101