fix: open URLs on Windows (#116)
The URLs on Windows failed to be opened because the cmd executed via
ProcessExecutor was not correctly constructed. We were calling
`exec("cmd", "start \"$url\"")` but in Windows `/c` is also needed to
the `cmd`.

We originally used native commands to open URLs because Toolbox didn’t
support it. Now that LocalDesktopManager provides an API for launching
the browser, we no longer need to fix the command-line logic — we can
just use the Toolbox API instead.