fix: url refresh after switching deployments
The main env header page API is quite limiting, in the sense that the title is never allowed to change.
Today we display the Coder URL as the title. However, if the user switches between two deployments via
log out and then log in, the URL is never refreshed, leading to a confusing UI (URL is old, while workspaces
are from the new deployment)

JetBrains suggested a workaround in https://youtrack.jetbrains.com/issue/TBX-14421/ by creating a new env
page instance with the updated title after each deployment switch. In which case Toolbox redraws the main
screen when page state changes.
I tested the workaround and indeed it works but with a caveat. The new env page is refreshed only when
Toolbox window page is made visible or when navigating from main env page to a single env page and then
back to the main page. For some reason, if we call `envPageManager.showPluginEnvironmentsPage()`` from the
authentication page displayed in `getOverrideUiPage() it doesn't trigger the new env refresh giving the
false impression that the URL won't change. In fact this was the reason why I missed this workaround in previously.

It is interesting that if I first navigate to a blank page - `toolboxUi.showUiPage(emptyPage)` - and then
immediately call `envPageManager.showPluginEnvironmentsPage()`, doing this rapid sequence seems to do the
trick without any visible artifacts (it happens so fast that we can't see the blank page being rendered)

- resolves #66