fix: remember token when switching deployments
If we log in on deployment 1, then log out and login to deployment 2 and then in the same session
we try to log in back to deployment 1, the token is no longer valid. The plugin will associate with
deployment 1 the token from the second deployment.

There is an overly complicated block of code inherited from Gateway plugin with multiple fallback sequences
for both the deployment url and token from multiple sources (secrets store, data dir config, env, etc...).
This fix simplifies the approach, we only store the url and the token in the secrets store, the token is
always associated to a hostname. If there is no previous URL to remember (like the first time login) we
default to https://dev.coder.com/ and empty token.