chore: simplify uri handling implementation (1)
Refactored code around uri, token, workspace and rest client resolving by
encapsulating code in clearly named methods. Rest client resolving was overly-complicated
(code inherited from Gateway), with token being a mandatory parameter. Removed a lot of code
that asked the token from the user if it was missing.

Also, I decided to use a snackbar to show errors because of attached regression in TBX.
A bottom simple dialog is more pleasing to the eye than a dialog in the middle of a page.

https://youtrack.jetbrains.com/issue/TBX-14944/Pop-up-dialogs-are-only-displayed-on-the-main-envs-page