fix: show errors when TBX is visible after being minimized
Errors encountered while TBX was running but the window was not visible were never
displayed by TBX. This fix queues the errors while TBX is minimized, and they will be
displayed again only when visible. This implementation is possible due to an observable
state object that can provide information about TBX and plugin visibility.

Among other things we also display a more human friendly version for the exceptions
raised by the http client during (but not only) workspace polling.