impl: support for Toolbox 2.6.3 (#124)
Toolbox 2.6.3 comes with a couple of new additions in the API which need
the following changes:

- finish support for URI handling. The available API up to TBX 2.6.3 was
buggy in terms of URI handling. It didn't allow plugins to
programmatically install remote ides and launch them. The launch
operation only worked when the IDE was already installed and a project
was already opened with the IDE.

TBX 2.6.3 adds a new API, _RemoteToolboxHelp_ which provides routines
for listing the available
IDEs on the remote, what is already installed and a command to install
specific versions of the IDE.
Additionally, there were fixes provided to the existing _ClientHelper_
which now launches the JBClient
if a project was not specified. An additional quirk I've discovered is
that if we provide a project, and
that project was not already opened (present in the Projects tab) the
IDE still won't open. And there
  is no API available to query the available projects.
  
  This PR uses the new API to:
  - query the installed ides
  - check if the provided ide is in the list of already installed IDEs.
- if that's not the case we query the available list of IDEs and the
available versions
- if the provided ide and build no., is in the available list we will
schedule it for install
- if not, we select the latest available build number for the provided
product code.
  - wait for the remote IDE to be installed
- and then download and launch the JBClient with a project path if it
was provided.
- update the minimum API requirement. Toolbox API is upgraded to
1.1.41749 which comes with new API additions and some deprecations.
Kotlin stdlib was also increased to a newer patch version

- use new environment state API. The _CustomRemoteEnvironmentState_ is
deprecated, and replaced by a new class _CustomRemoteEnvironmentStateV2_
which now supports i18n state labels

- use the new ssh disconnect callback. Toolbox provides two callbacks,
one before an SSH connection is established and another one which
executes when the ssh connection is stopped. The latter was deprecated
in the favor of a new callback that also provides hints on whether the
user requested the disconnect.

- use the new delete callback API. Toolbox provides a callback for
scenarios that involve the env. deletion. This allows plugins to react
and clean the internal state. With the new TBX API, the delete callback
API is deprecated in the favor of a mutable state flow, a reactive
approach that allows consumers to observe and react to state changes
over time.