fix: update&start outdated workspaces (#128)
URI handling was not able to start workspaces that were stopped and also
outdated. With this PR we check the ws status and call the proper REST
endpoint depending on whether the workspace is outdated or not.

During URI handling we can wait for workspaces to start up. The existing
implementation had no visual feedback that we are waiting/doing
anything. With this PR the header bar shows a nice work in progress
visual. Additionally, the deployment URL was not properly refreshed when
switching between different urls via URI handling. This is also fixed by
forcing TBX to render a blank page for a very short period of time and
then going back to the main page.