impl: remember the ssh connection state (#125)
And try to automatically establish the connections after an expired
token was refreshed (by going again through the login sequence)

In addition a fix was provided in order to show errors when TBX is
visible after being minimized. Errors encountered while TBX was running
but the window was not visible were never displayed by TBX. This fix
queues the errors while TBX is minimized, and they will be displayed
again only when visible. This implementation is possible due to an
observable
state object that can provide information about TBX and plugin
visibility.

Among other things we also display a more human friendly version for the
exceptions raised by the http client during (but not only) workspace
polling.

Attention: users will still have to manually launch a new a remote IDE
if it was opened while a session expired.

- resolves #121