impl: prioritize workspace status over TBX agent status
TBX 2.7 API prioritizes connection status to the remote TBX agent over
the workspace status. This would mean that a running workspace would be
listed as disconnected instead of ready. This is confusing as it only
happens for running workspaces and on top of that we also have a green
badge that is rendered by TBX when the agent connection is established.