chore: remove attention note related to the `folder` query param
Toolbox 2.7 allows the ide to open paths (via uri handling) on the remote that
were not previously a proper IDEA project.

There is a minor issue that I raised in https://youtrack.jetbrains.com/issue/TBX-14952/ with
paths that point out to folders or files that do not exist. In this case, TBX pops up a dialog
saying "Connecting to remote host..."  and it stays like that forever.