fix: UTs after fallback to signatures from releases.coder.com were published
We now have a lot of signatures published for a lot of older cli version which means
some of the tests that were not expecting the fallback to activate are now in trouble.
The simple fix is to "download" a very old version for which signatures will not be generated.