impl: download the cli to a temporary location
The download and signature verification steps are now slightly altered to first download
the cli to a temporary location and only after the signature verification is successful
or the user accepted the risk of running an unsigned binary - then and only then the temp
cli is moved to its final location (actually it is just a rename).

If the delete fails, this prevents the unsigned binary from being picked up as the
cached binary on the next run.