impl: introduce signature fallback setting
We are going to ask the user only once during the initial login screen if he wants to
fallback on releases.coder.com when signatures are not present in the coder deployment.
However, we still want to leave him the option to later change his mind the in Settings page.