fix: the settings page doesn't see changes done from other screens
Unless we restart toolbox. This happens because the UI fields are initialized only
once when the page is created. If we switch to a different page that can update the
settings (for example the login screen can alter the fallback strategy when signature
are missing) and then come back to the settings page - in that case Toolbox does not
request the settings again from the settings store.