impl: verify cli signature (#148)
This PR introduces support for verifying the CLI binary using a detached
PGP signature. Starting with version 2.24, Coder signs all CLI binaries.
For clients using older versions or running TBX in air-gapped
environments, unsigned CLIs can still be executed — but users will have
to confirm it each time.

In terms of code changes - the PR includes a big refactor around CLI
downloading with most of the code refactored and extracted in various
components that provide clean steps and result state in the main
download method. Then the pgp verification logic was added on top, with
some particularities:
- the pgp public key is embedded in the plugin as a jar resource
- we support multiple key rings in the public key
- the user has the option of running the CLI if no signature was found
- the signature search has a fallback approach: first we look in the
Coder deployment, and then fall back to releases.coder.com to search for
the signature if the user allows it.
- we expect the signature to be under the same relative path as the CLI
(we have an option which allows user to pick the CLI from a different
source other than the Coder deployment)