impl: support for certificate based authentication (#155)
We now skip token input screen if the user provided a public and a
private key for mTLS authentication on both the usual welcome screen and
in the URI handling.

Attention: the official coder deployment supports only authentication
via token, which is why I could not fully test an end to end scenario.