impl: enhanced workflow for network disruptions (#162)
Currently, when the network connection drops, the Coder TBX plugin
resets itself, redirects users to the authentication page, and
terminates active SSH sessions to remote IDEs. This disrupts the user
experience, forcing users to manually reconnect once the network is
restored. Additionally, since the SSH session to the remote IDE is lost,
the JBClient is unable to re-establish a connection with the remote
backend.
    
This PR aims to improve that experience by adopting a behavior similar
to the SSH plugin. Instead of clearing the list of workspaces or
dropping existing SSH sessions during a network outage, we retain them.
Once the network is restored, the plugin will automatically reinitialize
the HTTP client and regenerate the SSH configuration—only if the number
of workspaces has changed during the disconnection—without requiring
user intervention.

Additionally we also add support for remembering SSH connections that
were not manually disconnected by the user. This allows the plugin to
automatically restore those connections on the next startup enabling
remote IDEs that remained open to reconnect once the SSH link is
re-established.