impl: hide configurable fallback if signature verification is disabled
The main plugin screen has a quick shortcut for setting whether the user
wants to fallback on releases.coder.com for signatures if they are not provided by
the main deployment. This checkbox should not be visible if the user wants to disable
signature verification altogether.

To make things more consistent with Coder Gateway, the fallback setting is always displayed
if signature verification is enabled, we no longer display it only once in the main page.