impl: poll workspaces when Toolbox screen becomes visible
Some users complained that Coder Toolbox has a noticeable delay in rendering the workspaces
status compared with the web Dashboard, even though they have the same polling frequency,
which is every 5 seconds.
However, the web Dashboard also reacts and when the tab receives focus from the user which
improves the experience. This PR tries to implement the same behavior, when Coder Toolbox
screen becomes visible the polling is triggered immediately which should result in a status update.