impl: report progress while handling URI
Up until now there was no progress while downloading CLI, setting up the cli and the ssh config
while handling URIs. This PR reworks the uri handler and the connection screen to be able to reuse
the later part in the URI handler. This should improve the experience because the user is no longer
left in the dark for a good couple of seconds.