fix: NPE during error reporting (#186)
The try/catch block raised NPE in the `notify` if another exception was
raised after the context containing the URL was reset - so that means an
error in the onConnect handler. In addition, some of the reset steps
were moved after onConnect to make sure they execute only if onConnect
callback is successful.

Because of the fault in how the steps were arranged, the original
exception was never logged instead a misleading NPE was treated by the
coroutine's exception handler.