fix: don't store token when certificates are configured (#192)
This was a regression for the custom flows that required authentication
via certificates, the http client and the coder cli were properly
initialized but afterward the token was still required for storing.

Note: the issue was hard to catch early on because the official coder
cli is not supporting auth. via certificates.