fix: report errors while running actions
JetBrains team reported in the past a couple of errors in the log, one of them
being `A workspace build is already active`. The issue can be reproduced if the user
hits the `Stop` action for example quite quick. It takes maybe one or two seconds
to make rest api request, then for the backend to enqueue the build and change the
workspace action. If we hit the action buttons really fast then this error could be
reproduced.

One approach I tried was to disable the action buttons in the context menu for the duration
the request is executed. But for some reason the "enabled" property is not working in context
menu, only when the actions are rendered on a UI "page".

Instead, I decided to refactor the existing code and (also) visually report the errors in the
UI screen to make the user aware in some cases that a job is already running on the backend.