impl: improve logging around resource cleanup
Another error reported by JetBrains is a  `RejectedExecutionException` in the rest api client,
and from the stack trace it seems the thread pool in the rest client was at some point shutdown.
I think it is some sort of race condition, some thread calling shutting down the rest api client
while the UI thread still executes polling and user's action. I tried to reproduce the issue with
no success, and so I'm improving the logging around plugin de-initialization in the hope that next
time the sequence of events is more helpful.