fix: report errors while running actions (#193)
JetBrains team reported in the past a couple of errors in the log, one
of them being `A workspace build is already active`. The issue can be
reproduced if the user hits the `Stop` action for example quite quick.
It takes maybe one or two seconds to make rest api request, then for the
backend to enqueue the build and change the workspace action. If we hit
the action buttons really fast then this error could be reproduced.

One approach I tried was to disable the action buttons in the context
menu for the duration the request is executed. But for some reason the
"enabled" property is not working in context menu, only when the actions
are rendered on a UI "page".

Instead, I decided to refactor the existing code and (also) visually
report the errors in the UI screen to make the user aware in some cases
that a job is already running on the backend.

Another error reported by JetBrains is a `RejectedExecutionException` in
the rest api client, and from the stack trace it seems the thread pool
in the rest client was at some point shutdown.
I think it is some sort of race condition, some thread calling shutting
down the rest api client while the UI thread still executes polling and
user's action. I tried to reproduce the issue with no success, and so
I'm improving the logging around plugin de-initialization in the hope
that next time the sequence of events is more helpful.