impl: confirmation dialog for workspace deletion (#179)
Users are now required to confirm the workspace name if they want to
delete a workspace. This is in order to avoid any accidental removals.

Note: right now there are two issues with Toolbox input dialogs, the
dialog title is not rendered, and worse - the text field is rendered as
a password input field, so it does not make sense to merge this until
Toolbox fixes the issues.

<img width="486" height="746" alt="image"
src="https://github.com/user-attachments/assets/e8fc2409-6e24-42d0-8258-473c29f5df44"
/>


- resolves #178