impl: ability to customize the link to workspace creation in Dashboard
Some clients (Netflix in this specific case) rely on mainly their own dashboard
tools instead of the Coder one. Two main reasons that were mentioned by Netflix:
- aggregate many dev tools in a unified internal console
- specific platform/security needs that their own UI handles better

For this reason they would like the actions open up the Coder Dashboard
(only Create workspace for now) to be fully customizable, and allow clients
to override the URL.
For `Create workspace` we now have a config that defaults $lastDeploymentUrl/templates,
but it can be replaced with a relative URL to $lastDeploymentUrl or a complete new URL.
For now the decision is to not allow configuration from UI since Netflix is the
only target for this change, and they deploy at scale a templated settings.json.