impl: ability to customize the links to Dashboard (#211)
Some clients (Netflix in this specific case) rely on mainly their own
dashboard tools instead of the Coder one. Two main reasons that were
mentioned by Netflix:
- aggregate many dev tools in a unified internal console
- specific platform/security needs that their own UI handles better

For this reason they would like the actions that open up the Coder
Dashboard (`Create workspace` and `Open in dashboard`) to be fully
customizable, and allow clients to override the URL.
For `Create workspace` we now have a config that defaults
$lastDeploymentUrl/templates, but it can be replaced with a complete new
URL. It also supports `$workspaceOwner` as a placeholder that is
replaced by the plugin with the username that logged in.

For `Open in dashboard` a full URL can be provided and we also
introduced two placeholders `$workspaceOwner` and `$workspaceName` which
will be replaced by the plugin but only for this action.

For now the decision is to not allow configuration from UI since Netflix
is the only target for this change, and they deploy at scale a templated
settings.json.