impl: use the latest build number available
The URI handler expects a product code a build number. If the build number
is no longer available we defaulted to latest version available and popped
up a confirmation window to make sure the user wants to go forward with
the new version

Netflix pointed out that the confirmation dialog disrupts the workflow.
With this PR we only log that we fall back to latest version available
in order to streamline the handler.