fix: show connect page in the URI handler
If Toolbox is already opened and URI is executed the setup page
from below is never called. I tried a couple of things, including
yielding the coroutine - but it seems to be of no help. What works
delaying the coroutine for 66 - to 100 milliseconds, these numbers
were determined by trial and error.
The only explanation that I have is that inspecting the TBX bytecode it seems the
UI event is emitted via MutableSharedFlow(replay = 0) which has a buffer of 4 events
and a drop oldest strategy. For some reason it seems that the UI collector
is not yet active, causing the event to be lost unless we wait > 66 ms.
I think this delay ensures the collector is ready before processEvent() is called.