impl: ability to application name as main page title (#220)
Netflix would like the ability to use application name displayed in the
dashboard as the main page title instead of the URL.

This PR adds a new option `useAppNameAsTitle` that allows users to
specify whether or not they want to use the application name visible in
the dashboard as Tbx main tile instead of the URL. The default will
remain the URL.
Unlike previous settings added for Netflix this one is also configurable
from the UI (Coder Settings page) so not only via settings.json file.
This is an option that probably makes sense for more users.