impl: start the workspace via Coder CLI
Netflix uses custom MFA that requires CLI middleware to handle auth flow. The custom CLI
implementation on their side intercepts 403 responses from the REST API, handles the MFA
challenge, and retries the rest call again. The MFA challenge is handled only by the `start`
and `ssh` actions. The remaining actions can go directly to the REST endpoints because
of the custom header command that provides MFA tokens to the http calls.

Both Gateway and VS Code extension delegate the start logic to the CLI, but not Toolbox
which caused issues for the customer. This PR ports some of the work from Gateway in Coder Toolbox.