refactor: simplify workspace start status management
Current approach with a secondary poll loop that handles the start action
of a workspace is overengineered. Basically the problem is the CLI takes too
long before moving the workspace into the queued/starting state, during which
the user doesn't have any feedback. To address the issue we:
- stopped the main poll loop from updating the environment
- moved the environment in the queued state immediately after the start button was pushed.
- started a poll loop that moved the workspace from queued state to starting space only
after that state became available in the backend. The intermediary stopped state is skipped by
the secondary poll loop.

@asher pointed out that a better approach can be implemented. We already store the status, and workspace
and the agent in the environment. When the start comes in:
1. We directly update the env. status to "queued"
2. We only change the environment status if there is difference in the existing workspace&agent
status vs the status from the main poll loop
3. no secondary poll loop is needed.