fix: compare builds instead of statuses
when updating the workspace and agent. Checking the status is
subtly different from checking if there is a new build in the following scenario:

1. Latest build is failed.
2. The poll runs, so we currently have failed stored as the status.
3. User starts workspace, we set to queued.
4. Build fails again (a brand-new build, but it is also failing).
5. Poll runs again, checks failed against the last status, which is also failed,
so it leaves the status as queued, even though the workspace is actually failed
currently, so really we do want to update the state to failed.

So we could miss updates from one build to another build when both builds end up with the same status.